requirements.txt
google-genai>=1.47.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selenium>=4.18.0
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.5
Pillow>=10.0.0
python-dotenv>=1.0.0
